Rename middleware -> serviceFunction
